Separating out client interface test data
Separating out the reflection used to locate all API interfaces so that it can be reused with new tests
